extract utils and hook out of component file
